+++ NEWS.md
@@ -3,9 +3,18 @@ title: "NEWS"
 output: html_document
 ---
 
-```{r setup, include=FALSE}
-knitr::opts_chunk$set(echo = TRUE)
-```
+
+## backbone 1.2.0
+
+* add support for sparse matrix, igraph, network, and edgelist objects (see 'class.convert')
+* add family-wise error rate test corrections (see 'backbone.extract')
+* sdsm: add multiple methods for computing initial probabilities (see 'sdsm' details) one of which uses convex optimization (see 'polytope')
+* sdsm: update poisson binomial computation method to increase speed (see 'sdsm' and 'rna')
+* add more descriptives to summary dataframe output of backbone object
+* update documentation of functions
+* update vignette to reflect package changes
+* bug fixes for R 4.0.0
+
 ## backbone 1.1.0
 
 * add support for sparse matrices